update: adding logging data in api with excel, adding example question, and update information in README
